streams: gradle multi-project build
