+++ chunk-processing-api/src/main/java/org/asheldo/streams/chunkproc/api/model/FeedInputBase.java
@@ -1,0 +1,21 @@
+package org.asheldo.streams.chunkproc.api.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class FeedInputBase {
+
+    @JsonProperty
+    @Getter
+    private String s3Prefix;
+
+    @JsonProperty
+    @Getter
+    private String channel;
+
+    @JsonProperty
+    @Getter
+    private FeedBatchStepConfig feedBatchStepConfig;
+}
