+++ chunk-spliterator-lib/build.gradle
@@ -1,0 +1,45 @@
+plugins {
+    id 'java'
+    id 'java-library'
+    id 'groovy'
+    id 'codenarc'
+    id "io.freefair.lombok" version "4.1.5"
+}
+
+group 'org.asheldo'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+codenarcMain {
+    configFile = rootProject.file("codenarc/rules.groovy")
+}
+
+codenarcTest {
+    configFile = rootProject.file("codenarc/rules-test.groovy")
+}
+
+dependencies {
+    implementation(
+            project(':chunk-processing-api'),
+            'ch.qos.logback:logback-classic:1.2.3',
+            'com.google.guava:guava:28.1-jre',
+            'com.google.inject:guice:4.2.2',
+            'com.fasterxml.jackson.core:jackson-annotations:2.10.1',
+            'com.fasterxml.jackson.core:jackson-core:2.10.1',
+            'com.fasterxml.jackson.core:jackson-databind:2.10.1',
+            'software.amazon.awssdk:bom:2.10.30',
+            'software.amazon.awssdk:s3:2.10.30',
+            'commons-io:commons-io:2.6'
+    )
+    testImplementation(
+            'junit:junit:4.12',
+            'org.codehaus.groovy:groovy-all:2.5.8',
+            'org.spockframework:spock-core:1.3-groovy-2.5'
+    )
+}
