+++ chunk-spliterator-lib/src/main/java/org/asheldo/streams/RemoteOutputs.java
