+++ chunk-spliterator-lib/src/main/java/org/asheldo/streams/partition/StringWithIndex.java
