+++ chunk-spliterator-lib/src/test/groovy/org/asheldo/streams/LocalOutputsSpec.groovy
