--- src/main/java/org/asheldo/streams/FeedBatchStepSettings.java
@@ -1,5 +1,0 @@
-package org.asheldo.streams;
-
-public interface FeedBatchStepSettings {
-    String getChannel();
-}
