--- src/main/java/org/asheldo/streams/LastExposureSettingsImpl.java
@@ -1,18 +1,0 @@
-package org.asheldo.streams;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@AllArgsConstructor
-public class LastExposureSettingsImpl implements FeedBatchStepSettings {
-
-    @JsonProperty
-    @Getter
-    @Setter
-    private String channel;
-
-}
